Remove redundant TextUtility class
